Recent posts migrated to styled components :rocket:
