+++ src/styles/blog-post.scss
@@ -1,6 +1,4 @@
 .blog-post {
-  margin-bottom: 50px;
-
   ul {
     li{
       font-family: 'Open Sans', Helvetica, Arial, sans-serif;
@@ -9,7 +7,6 @@
 
   figure {
     figcaption {
-      @include theme('color', $secondary-text-color);
       font-size: 14px;
       text-align: center;
     }
