--- src/styles/blog-recent-posts.scss
@@ -1,57 +1,0 @@
-.recent-posts {
-  margin-bottom: 35px;
-
-  .row{
-    margin-top:20px;
-  }
-
-  a {
-    @include theme('color', $primary-color-text);
-    &:hover{
-      text-decoration:none;
-    }
-  }
-
-  .cards {
-    display: flex;
-    width: 100%;
-    @include media-breakpoint-down(sm) {
-      display: block;
-    }
-
-    h3 {
-      @include theme('color', $accent-color);
-      font-weight: 500;
-      @include media-breakpoint-down(sm) {
-        padding-left: 0;
-      }
-    }
-
-    @include media-breakpoint-down(sm) {
-      .col-lg-4 {
-        margin-top: 20px;
-      }
-    }
-
-    .card {
-      @include theme('background-color', $general-background-light);
-      @include theme('box-shadow', $box-shadow-light, true, '0 3px 10px 0');
-      padding: 0 0 80px 0;
-      width: 100%;
-      height: 100%;
-      border-radius:4px;
-      border: none;
-      position: relative;
-      @include media-breakpoint-down(sm) {
-        margin: 10px 0 0 0;
-      }
-
-      @include media-breakpoint-up(lg) {
-        transition: all .2s;
-        &:hover {
-          transform: scale(1.025);
-        }
-      }
-    }
-  }
-}
