+++ src/styles/style.blog.post.scss
@@ -2,4 +2,3 @@
 @import "./common.scss";
 @import "./blog-common.scss";
 @import "./blog-post.scss";
-@import "./blog-recent-posts.scss";
